+++ web/src/components/trace/TracePreview.tsx
@@ -74,16 +74,6 @@ export const TracePreview = ({
   const showScoresTab = isAuthenticatedAndProjectMember && peek === undefined;
   const { expansionState, setFieldExpansion } = useJsonExpansion();
 
-  const inputKeys = Object.keys(expansionState.input).filter(
-    (k) => expansionState.input[k],
-  ).length;
-  const outputKeys = Object.keys(expansionState.output).filter(
-    (k) => expansionState.output[k],
-  ).length;
-  const metadataKeys = Object.keys(expansionState.metadata).filter(
-    (k) => expansionState.metadata[k],
-  ).length;
-
   const traceMedia = api.media.getByTraceOrObservationId.useQuery(
     {
       traceId: trace.id,
